Add the msysgit package for mswindows.
This package has some problems. It's not just Git but someting like
a sysroot. Then this package should not be installed directly in
devroot/. It should be installed in devroot/opt/msysgit/.

The question is: "How to update the PATH accordingly to the package
specificities?". It suggests that the answer is with the package
definition file but it's an other issue.