Convert microsoft+buildtools-2019 to microsoft+buildtools
The main reason is that it's not really possible to have
multiple buildtools versions side by side. If you think
that is supported, you are wrong because the installer
must always be the last. It's just partially supported
when using the Visual Studio Installer GUI but in practice
it's just bullshit.